[mds-agency] Maxj/mds agency tests (#177)
* WIP

* refactor endpoint handlers

* further cleanup

* add first test!

* more tests+refactors

* I am a bit confused

* progress!!!

* split up error handling

* add another test

* fix type error

* add more tests and temp disable nyc threshold

* recheck in old tests
