+++ packages/mds-agency/types.ts
@@ -1,4 +1,4 @@
-import { UUID } from '@mds-core/mds-types'
+import { UUID, Device, VehicleEvent, Telemetry } from '@mds-core/mds-types'
 import { MultiPolygon } from 'geojson'
 import { ApiRequest, ApiResponse, ApiResponseLocals } from '@mds-core/mds-api-server'
 
@@ -19,3 +19,19 @@ export interface ServiceArea {
   description: string
   area: MultiPolygon
 }
+
+export interface VehiclePayload {
+  device?: Device
+  event?: VehicleEvent
+  telemetry?: Telemetry
+}
+
+export type TelemetryResult =
+  | Telemetry
+  | Readonly<
+      Required<
+        Telemetry & {
+          id: number
+        }
+      >
+    >[]
